requirements.txt
pypdf
pandas
pytesseract
pdf2image
Pillow
